+++ public/licenses.json
@@ -1,1 +1,1 @@
-[{"name":"@fortawesome/fontawesome-svg-core","link":"git+https://github.com/FortAwesome/Font-Awesome.git","installedVersion":"6.5.1","author":"The Font Awesome Team (https://github.com/orgs/FortAwesome/people)"},{"name":"@fortawesome/free-brands-svg-icons","link":"git+https://github.com/FortAwesome/Font-Awesome.git","installedVersion":"6.5.1","author":"The Font Awesome Team (https://github.com/orgs/FortAwesome/people)"},{"name":"@fortawesome/free-regular-svg-icons","link":"git+https://github.com/FortAwesome/Font-Awesome.git","installedVersion":"6.5.1","author":"The Font Awesome Team (https://github.com/orgs/FortAwesome/people)"},{"name":"@fortawesome/free-solid-svg-icons","link":"git+https://github.com/FortAwesome/Font-Awesome.git","installedVersion":"6.5.1","author":"The Font Awesome Team (https://github.com/orgs/FortAwesome/people)"},{"name":"@fortawesome/react-fontawesome","link":"git+https://github.com/FortAwesome/react-fontawesome.git","installedVersion":"0.2.0","author":"n/a"},{"name":"bootstrap","link":"git+https://github.com/twbs/bootstrap.git","installedVersion":"5.3.2","author":"The Bootstrap Authors (https://github.com/twbs/bootstrap/graphs/contributors)"},{"name":"next","link":"git+https://github.com/vercel/next.js.git","installedVersion":"14.1.0","author":"n/a"},{"name":"pdf2pic","link":"git+https://github.com/yakovmeister/pdf2image.git","installedVersion":"3.1.1","author":"Jacob Baring <owari@waifu.club>"},{"name":"react","link":"git+https://github.com/facebook/react.git","installedVersion":"18.2.0","author":"n/a"},{"name":"react-bootstrap","link":"git+https://github.com/react-bootstrap/react-bootstrap.git","installedVersion":"2.10.0","author":"Stephen J. Collings stevoland@gmail.com"},{"name":"react-dom","link":"git+https://github.com/facebook/react.git","installedVersion":"18.2.0","author":"n/a"}]
+[{"name":"@fortawesome/fontawesome-svg-core","link":"git+https://github.com/FortAwesome/Font-Awesome.git","installedVersion":"6.5.1","author":"The Font Awesome Team (https://github.com/orgs/FortAwesome/people)"},{"name":"@fortawesome/free-brands-svg-icons","link":"git+https://github.com/FortAwesome/Font-Awesome.git","installedVersion":"6.5.1","author":"The Font Awesome Team (https://github.com/orgs/FortAwesome/people)"},{"name":"@fortawesome/free-regular-svg-icons","link":"git+https://github.com/FortAwesome/Font-Awesome.git","installedVersion":"6.5.1","author":"The Font Awesome Team (https://github.com/orgs/FortAwesome/people)"},{"name":"@fortawesome/free-solid-svg-icons","link":"git+https://github.com/FortAwesome/Font-Awesome.git","installedVersion":"6.5.1","author":"The Font Awesome Team (https://github.com/orgs/FortAwesome/people)"},{"name":"@fortawesome/react-fontawesome","link":"git+https://github.com/FortAwesome/react-fontawesome.git","installedVersion":"0.2.0","author":"n/a"},{"name":"axios","link":"git+https://github.com/axios/axios.git","installedVersion":"1.6.5","author":"Matt Zabriskie"},{"name":"bootstrap","link":"git+https://github.com/twbs/bootstrap.git","installedVersion":"5.3.2","author":"The Bootstrap Authors (https://github.com/twbs/bootstrap/graphs/contributors)"},{"name":"next","link":"git+https://github.com/vercel/next.js.git","installedVersion":"14.1.0","author":"n/a"},{"name":"pdf2pic","link":"git+https://github.com/yakovmeister/pdf2image.git","installedVersion":"3.1.1","author":"Jacob Baring <owari@waifu.club>"},{"name":"react","link":"git+https://github.com/facebook/react.git","installedVersion":"18.2.0","author":"n/a"},{"name":"react-bootstrap","link":"git+https://github.com/react-bootstrap/react-bootstrap.git","installedVersion":"2.10.0","author":"Stephen J. Collings stevoland@gmail.com"},{"name":"react-dom","link":"git+https://github.com/facebook/react.git","installedVersion":"18.2.0","author":"n/a"}]
