+++ src/app/amtsblaetter/page.js
@@ -1,9 +1,12 @@
-
 import Header from "@/app/components/Header";
+import fs from "fs/promises";
+import path from "path";
+import { fromBuffer } from "pdf2pic";
 import { Container } from "react-bootstrap";
 import Col from "react-bootstrap/Col";
 import Row from "react-bootstrap/Row";
 import Amtsblatt from "../components/Amtsblatt";
+
 
 const Amtsblaetter = [
     [
@@ -19,8 +22,30 @@ const Amtsblaetter = [
 ];
 
 async function fetchPreviews() {
-    const res = await fetch("https://127.0.0.1:3000/api/imgtopdf")
-    const { images } = await res.json()
+    const basePath = path.resolve(process.cwd() + "/public/amtsblaetter");
+    async function getPdfImages(filePath) {
+        const options = {
+            density: 100,
+            format: "png",
+            width: 1080,
+            height: 1920,
+        };
+        const file = await fs.readFile(path.resolve(filePath));
+        const convert = fromBuffer(file, options);
+        const pageToConvertAsImage = 1;
+        const meta = await convert(pageToConvertAsImage, {
+            responseType: "base64",
+        });
+        return meta.base64;
+    }
+
+    const blaetter = (await fs.readdir(basePath)).sort();
+    const images = [];
+    for (const doc of blaetter) {
+        const img = await getPdfImages(path.resolve(basePath + `/${doc}`));
+        images.push(img);
+    }
+
     return images.map((image) => `data:image/png;base64,${image}`);
 }
